Refactor onboarding and correct constrains for flexible tracks
Change-type: patch
