+++ helper-scripts/download-and-configure-input.js
@@ -23,65 +23,32 @@ const { getNextCycleDates } = require('../lib/gsheets');
 const { getSchedulerInput } = require('../lib/gsheets');
 const { validateJSONScheduleInput } = require('../lib/validate-json');
 
-// The following constanst are for configuration porpouses and should be left empty if not needed
-const specialAgentConditions = {
-	agentsWithMaxHoursShift: [
-		{
-			handle: '@samothx',
-			value: 4,
-		},
-	],
-	agentsWithMinHoursWeek: [
-		{
-			handle: '@saintaardvark',
-			value: 7,
-		},
-	],
-	agentsWithFixHours: ['@georgiats'],
-};
-
-// TRACKS to be used in the calendar.
-// Tracks are parallel shifts ocurring on the same day (3 agents at the same time need three tracks).
-// If shifts need to be scheduled after midnight
-const tracks = [
-	{ start_hour: 6, end_hour: 26, start_day: 0, end_day: 4 },
-	{ start_hour: 6, end_hour: 26, start_day: 0, end_day: 4 },
-	{ start_hour: 6, end_hour: 12, start_day: 0, end_day: 0 },
-];
-
-const SCHEDULE_OPTS = {
-	numConsecutiveDays: 5,
-	numSimultaneousTracks: 2,
-	supportStartHour: 6,
-	supportEndHour: 26,
-	shiftMinDuration: 2,
-	shiftMaxDuration: 8,
-	optimizationTimeout: 3600,
-	specialAgentConditions,
-	tracks,
-};
-
 /**
  * Read and configure input data from Google Sheets, and save as JSON object
  */
-async function getData() {
+async function getData(supportModel) {
+	const scheduleOpts = JSON.parse(
+		fs.readFileSync('helper-scripts/options/' + supportModel + '.json')
+	);
+
 	try {
-		const auth = await getAuthClient();
+		const auth = await getAuthClient(supportModel);
 		const parsedNextCycleDates = await getNextCycleDates(auth);
 		const nextMondayDate = parsedNextCycleDates.support;
 
 		const schedulerInput = await getSchedulerInput(
 			auth,
 			nextMondayDate,
-			SCHEDULE_OPTS.numConsecutiveDays + 1
+			scheduleOpts.numConsecutiveDays + 1,
+			supportModel
 		);
-		_.assign(schedulerInput.options, SCHEDULE_OPTS);
+		_.assign(schedulerInput.options, scheduleOpts);
 		console.log(JSON.stringify(schedulerInput, null, 2));
 		const schedulerInputValidation = await validateJSONScheduleInput(
 			schedulerInput
 		);
 
-		const fileDir = `./logs/${nextMondayDate}`;
+		const fileDir = `./logs/${nextMondayDate}_` + supportModel;
 		await mkdirp(fileDir);
 
 		await fs.writeFile(
@@ -93,4 +60,14 @@ async function getData() {
 	}
 }
 
-getData();
+// Read scheduling algorithm output file name from command line:
+const args = process.argv.slice(2);
+if (args.length != 1) {
+	console.log(`please specify Scheduling Options file`);
+	process.exit(1);
+}
+
+// Load JSON object from output file:
+const supportModel = args[0];
+
+getData(supportModel);
