+++ helper-scripts/options/balena.json
@@ -1,0 +1,30 @@
+{
+  "modelName": "balena",
+  "numConsecutiveDays": 5,
+  "supportStartHour": 6,
+  "supportEndHour": 26,
+  "slotsInDay": 26,
+  "shiftMinDuration": 2,
+  "shiftMaxDuration": 8,
+  "optimizationTimeout": 3600,
+  "specialAgentConditions": {
+    "agentsWithMaxHoursShift": [
+      {
+        "handle": "@samothx",
+        "value": 4
+      }
+    ],
+    "agentsWithMinHoursWeek": [
+      {
+        "handle": "@saintaardvark",
+        "value": 7
+      }
+    ],
+    "agentsWithFixHours": ["@georgiats"]
+  },
+  "tracks": [
+    { "start_hour": 6, "end_hour": 26, "start_day": 0, "end_day": 4 },
+    { "start_hour": 6, "end_hour": 26, "start_day": 0, "end_day": 4 },
+    { "start_hour": 6, "end_hour": 12, "start_day": 0, "end_day": 0 }
+  ]
+}
