+++ helper-scripts/send-calendar-invites.js
@@ -35,12 +35,28 @@ async function readAndParseJSONSchedule(jsonPath) {
 	return jsonObject;
 }
 
+function getDate(eventDate, eventHour) {
+	resultDateTime = '';
+
+	if (eventHour > 23) {
+		let finalDate = new Date(Date.parse(eventDate));
+		finalDate.setDate(finalDate.getDate() + 1);
+		finalDate = finalDate.toISOString().split('T')[0];
+		endHour = eventHour - 24;
+		resultDateTime = `${finalDate}T0${endHour}:00:00`;
+	} else {
+		resultDateTime = `${eventDate}T${_.padStart(eventHour, 2, '0')}:00:00`;
+	}
+
+	return resultDateTime;
+}
+
 /**
  * From the object containing the optimized shifts, create array of "events resources" in the format required by the Google Calendar API.
  * @param  {object}   shiftsObject   Shifts optimized by scheduling algorithm
  * @return {array}                   Array of events resources to be passed to Google Calendar API.
  */
-async function createEventResourceArray(shiftsObject) {
+async function createEventResourceArray(shiftsObject, isProductOS) {
 	const returnArray = [];
 	for (const epoch of shiftsObject) {
 		const date = epoch.start_date;
@@ -49,30 +65,20 @@ async function createEventResourceArray(shiftsObject) {
 			let [handle, email] = shift.agent.split(' ');
 			email = email.match(new RegExp(/<(.*)>/))[1];
 
-			eventResource.summary = `${handle} on support`;
+			const supportName = isProductOS ? 'ProductOS support' : 'support';
+
+			eventResource.summary = `${handle} on ${supportName}`;
 			eventResource.description =
 				'Resources on support: ' + process.env.SUPPORT_RESOURCES;
 			eventResource.start = {
 				timeZone: TIMEZONE,
-				dateTime: `${date}T${_.padStart(shift.start, 2, '0')}:00:00`,
+				dateTime: getDate(date, shift.start),
 			};
 			eventResource.end = {
 				timeZone: TIMEZONE,
+				dateTime: getDate(date, shift.end),
 			};
 
-			if (shift.end > 23) {
-				let endDate = new Date(Date.parse(date));
-				endDate.setDate(endDate.getDate() + 1);
-				endDate = endDate.toISOString().split('T')[0];
-				endHour = 24 - shift.end;
-				eventResource.end.dateTime = `${endDate}T0${endHour}:00:00`;
-			} else {
-				eventResource.end.dateTime = `${date}T${_.padStart(
-					shift.end,
-					2,
-					'0'
-				)}:00:00`;
-			}
 			eventResource.attendees = [];
 			eventResource.attendees.push({ email: email });
 			returnArray.push(eventResource);
@@ -85,18 +91,26 @@ async function createEventResourceArray(shiftsObject) {
  * Load JSON object containing optimized schedule from file, and write to Support schedule Google Calendar, saving ID's of created events for reference.
  * @param  {string}   jsonPath   Path to JSON output of scheduling algorithm
  */
-async function createEvents(jsonPath) {
+async function createEvents(jsonPath, modelName) {
+	const isProductOS = modelName === 'productOS';
 	try {
 		const shiftsObject = await readAndParseJSONSchedule(jsonPath);
-		const eventResourceArray = await createEventResourceArray(shiftsObject);
-		const jwtClient = await getAuthClient();
+		const eventResourceArray = await createEventResourceArray(
+			shiftsObject,
+			isProductOS
+		);
+		const authClient = await getAuthClient(isProductOS ? 'token' : 'JWT');
 		const calendar = google.calendar({ version: 'v3' });
 		const eventIDs = [];
 
+		const calendarId = isProductOS
+			? process.env.PRODUCT_OS_CALENDAR_ID
+			: process.env.BALENA_CALENDAR_ID;
+
 		for (const eventResource of eventResourceArray) {
 			const eventResponse = await calendar.events.insert({
-				auth: jwtClient,
-				calendarId: process.env.CALENDAR_ID,
+				auth: authClient,
+				calendarId,
 				conferenceDataVersion: 1,
 				sendUpdates: 'all',
 				resource: eventResource,
@@ -120,13 +134,14 @@ async function createEvents(jsonPath) {
 
 // Read scheduler output file name from command line:
 const args = process.argv.slice(2);
-if (args.length != 1) {
+if (args.length != 2) {
 	console.log(
-		`Usage: node ${__filename} <path-to-support-shift-scheduler-output.json>`
+		`Usage: node ${__filename} <path-to-support-shift-scheduler-output.json> <model-name>`
 	);
 	process.exit(1);
 }
 const jsonPath = args[0];
+const modelName = args[1];
 
 // Derive path for output:
 let logsFolder = '';
@@ -137,4 +152,4 @@ if (jsonPath.indexOf('/') === -1) {
 }
 
 // Create calendar events:
-createEvents(jsonPath);
+createEvents(jsonPath, modelName);
