+++ lib/gauth.js
@@ -23,11 +23,21 @@ const SCOPES = [
 	'https://www.googleapis.com/auth/calendar',
 ];
 
+const TOKEN_PATH = process.env.TOKEN;
+const CRED = process.env.CREDENTIALS;
 /**
  * Obtains a Google OAuth 2.0 access token for access to Sheets and Calendar.
  * @return {object}     OAuth 2.0 access token
  */
-async function getAuthClient() {
+async function getAuthClient(authType) {
+	if (authType === 'JWT') {
+		return getJWTAuthClient();
+	} else {
+		return getOAuth2Client();
+	}
+}
+
+async function getJWTAuthClient() {
 	const content = await fs.readFile(process.env.GAPI_SERVICE_ACCOUNT_JWT);
 	const jwt = JSON.parse(content);
 	let auth = new google.auth.JWT({
@@ -37,6 +47,23 @@ async function getAuthClient() {
 	});
 	await auth.authorizeAsync();
 	return auth;
+}
+
+async function getOAuth2Client() {
+	const content = fs.readFileSync(CRED, 'utf8');
+	const oAuth = JSON.parse(content);
+	const { client_secret, client_id, redirect_uris } = oAuth.installed;
+	const oAuth2Client = new google.auth.OAuth2(
+		client_id,
+		client_secret,
+		redirect_uris[0]
+	);
+
+	const token = fs.readFileSync(TOKEN_PATH, 'utf-8');
+
+	oAuth2Client.setCredentials(JSON.parse(token));
+
+	return oAuth2Client;
 }
 
 exports.getAuthClient = getAuthClient;
