+++ lib/gsheets.js
@@ -172,6 +172,7 @@ async function parseInput(rawInput, startDate = null, numDays = 5) {
 		});
 		schedulerInput.agents.push(newAgent);
 	}
+	//TODO only if over 24 hours
 	schedulerInput.agents.forEach(agent => {
 		agent.availableHours.forEach((day, dayIndex) => {
 			let followingDay = dayIndex + 1;
@@ -193,11 +194,13 @@ async function parseInput(rawInput, startDate = null, numDays = 5) {
  * @param  {number} numDays        Number of consecutive days to schedule
  * @return {object}                Parsed input object for scheduler (more options will be added to this object by download-and-configure-input.js)
  */
-async function getSchedulerInput(auth, nextMondayDate, numDays) {
+async function getSchedulerInput(auth, nextMondayDate, numDays, supportModel) {
 	const sheets = google.sheets({ version: 'v4', auth });
+	const sheet = supportModel === 'productOS' ? '_productOS' : '_input';
+	const range = nextMondayDate + sheet + '!A3:KF';
 	const result = await sheets.spreadsheets.values.get({
 		spreadsheetId: process.env.SUPPORT_SCHEDULER_HISTORY_ID,
-		range: nextMondayDate + '_input!A3:KF',
+		range,
 		valueRenderOption: 'FORMATTED_VALUE',
 	});
 	const parsedInput = await parseInput(
